🚚️ api: move usecase to src/enrolment context
